refactor: common 패턴 따르도록 수정
