+++ src/users/users.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Users } from '../_entities/Users';
-import { UsersController } from './users.controller';
+import { Users } from '../common/entities/Users';
+import { UsersController } from './controller/users.controller';
 import { UsersRepository } from './users.repository';
-import { UsersService } from './users.service';
+import { UsersService } from './service/users.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Users])],
