+++ src/users/users.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Users } from '../_entities/Users';
+import { Users } from '../common/entities/Users';
 
 @Injectable()
 export class UsersRepository {
